gcp/requirements.txt
tensorflow== 2.13.0
google-cloud-storage==2.11.0
Pillow==10.0.1
pyttsx3==2.90


